added resuable cards to faq section of landing page
