+++ src/components/dashboard/styles.css
@@ -1,0 +1,3 @@
+#home {
+    width: 80%;
+}
