--- src/components/home/styles.css
@@ -1,3 +1,0 @@
-#home {
-    width: 80%;
-}
